refactor(efd-server): ♻️ filter world objects in WorldObjectProcessor
